+++ R/reexport.R
@@ -1,0 +1,3 @@
+#' @importFrom utils getFromNamespace
+#' @keywords internal
+geneticMapPlot <- getFromNamespace(x = "geneticMapPlot", ns = "statgenGWAS")