+++ R/singleGenoPlot.R
@@ -28,7 +28,7 @@ singleGenoPlot <- function(markers,
   p <- ggplot2::ggplot(plotDat,
                        ggplot2::aes(x = .data[["pos"]], y = .data[["parent"]],
                                     fill = .data[["prob"]])) +
-    ggplot2::geom_tile(width = 3) +
+    ggplot2::geom_tile(width = 0.03 * max(map$pos)) +
     ggplot2::facet_grid(". ~ chr", scales = "free", space = "free",
                         switch = "both") +
     ggplot2::scale_fill_gradient(low = "white", high = "black") +