+++ inst/doc/IBDFileFormat.R
@@ -23,7 +23,7 @@ ibdFile <- system.file("extdata/SxM", "SxM_IBDs.txt", package = "statgenIBD")
 ibds <- read.delim(ibdFile)
 knitr::kable(head(ibds))
 
-## ---- eval=FALSE------------------------------------------------------------------------
+## ----eval=FALSE-------------------------------------------------------------------------
 #  ## Write IBDs to file, set values <0.05 to zero and only print 3 decimals.
 #  writeIBDs(IBDprob = SxMIBD, outFile = tempfile(fileext = ".txt"),
 #            decimals = 3, minProb = 0.05)
@@ -36,6 +36,6 @@ SxMMap <- SxMIBD$map
 SxMIBD <- readIBDs("SxM-IBD.txt", map = SxMMap)
 summary(SxMIBD)
 
-## ---- echo=FALSE, results='hide'--------------------------------------------------------
+## ----echo=FALSE, results='hide'---------------------------------------------------------
 unlink("SxM-IBD.txt")