+++ NEWS.md
@@ -1,12 +1,18 @@
+# statgenIBD 1.0.6.9000
+
+* Plotting didn't work correctly for populations where there was a heterozygous probability in the IBD output. This is now corrected for plotType "allGeno", "meanProbs", and "totalCoverage".
+* It is now possible to create a plot of plotType "singleGeno" when the position of the markers in the map have large values, i.e. when they are in base pairs.
+* A small bug in the concatenation of `calcIBD` objects with an identical number of genotypes is fixed.
+
 # statgenIBD 1.0.6
 
-* A bug in the `plot` function is fixed. The "singlegeno" plots now show the correct probabilities. 
+* A bug in the `plot` function is fixed. The "singleGeno" plots now show the correct probabilities. 
 * Functions no longer rely on soft-deprecated ggplot2 functions.
 
 # statgenIBD 1.0.5
 
 * Two new functions for writing (`writeIBDs`) and reading IBD (`readIBDs`) probabilities to a custom format have been added. The format is explained in a new vignette `vignette("IBDFileFormat", package = "statgenIBD")`.
-* Two new plots are added for objects of class `IBDprob`, a plot showing the mean coverage of each parent across the population (`plotType = meanProbs`) and a plot showing the total coverage of each parent (`plotType = totalCoverage`).
+* Two new plots are added for objects of class `IBDprob`, a plot showing the mean coverage of each parent across the population (`plotType = "meanProbs"`) and a plot showing the total coverage of each parent (`plotType = "totalCoverage"`).
 * Parent names in the output of `calcIBD` no longer have a leading "p".
 * The function `readRabbit`, for reading output from the RABBIT software for computing IBDs for complex crosses, is moved here from `statgenMPP`.